requirements.txt
altair==4.2.2
attrs==25.1.0
audio-recorder-streamlit==0.0.10
audioop-lts==0.2.1
backports.functools-lru-cache==2.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
entrypoints==0.4
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
indic-transliteration==2.3.69
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
roman==5.0
rpds-py==0.23.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
SpeechRecognition==3.14.1
standard-aifc==3.13.0
standard-chunk==3.13.0
streamlit==1.43.1
tenacity==9.0.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
